+++ src/modules/tools/wfst/components/WfsTransaction.vue
@@ -9,7 +9,7 @@ export default {
     name: "WfsTransaction",
     components: {SimpleButton, ToolTemplate},
     computed: {
-        ...mapGetters("Tools/Wfst", ["currentInteractionConfig", "currentLayerIndex", "featureProperties", "layerIds", "layerInformation", "layerSelectDisabled", "layerSelectLabel", "selectedInteraction", "showInteractionsButtons", "active", "deactivateGFI", "icon", "name", "id"])
+        ...mapGetters("Tools/Wfst", ["currentInteractionConfig", "currentLayerIndex", "featureProperties", "layerIds", "layerInformation", "layerSelectDisabled", "layerSelectLabel", "selectedInteraction", "showInteractionsButtons", "active", "deactivateGFI", "icon", "name", "id", "isFormDisabled"])
     },
     watch: {
         active (val) {
@@ -39,7 +39,7 @@ export default {
     },
     methods: {
         ...mapMutations("Tools/Wfst", ["setCurrentLayerIndex", "setLayerInformation"]),
-        ...mapActions("Tools/Wfst", ["prepareInteraction", "reset", "save", "setActive", "setFeatureProperty", "setFeatureProperties"]),
+        ...mapActions("Tools/Wfst", ["prepareInteraction", "reset", "save", "setActive", "updateFeatureProperty", "setFeatureProperties"]),
         close () {
             this.setActive(false);
             const model = getComponent(this.id);
@@ -70,6 +70,31 @@ export default {
                 return "datetime-local";
             }
             return "";
+        },
+        getDecimalStep (type, value) {
+            if (type === "decimal" && value) {
+                const fractionalPartLength = value.toString().split(",")?.[1]?.length || value.toString().split(".")?.[1]?.length;
+
+                if (fractionalPartLength) {
+                    switch (fractionalPartLength) {
+                        case 1: return "0.1";
+                        case 2: return "0.01";
+                        default: return "0.001";
+                    }
+                }
+                return "1";
+            }
+            return null;
+        },
+        formatDecimalValue (type, value) {
+            if (type === "decimal" && value) {
+                if (value.includes(".") || value.includes(",")) {
+                    return value.replace(",", ".");
+                }
+                return value + ".0";
+
+            }
+            return value;
         }
     }
 };
@@ -138,6 +163,12 @@ export default {
                         >
                             {{ $t("common:modules.tools.wfsTransaction.polygonHint") }}
                         </p>
+                        <p
+                            v-if="featureProperties.find(prop => prop.required)"
+                            class="mb-2"
+                        >
+                            <span><span class="form-label-info"> - </span>{{ $t("common:modules.tools.wfsTransaction.fieldRequired") }}</span>
+                        </p>
                         <form id="tool-wfs-transaction-form">
                             <template v-for="property of featureProperties">
                                 <template v-if="property.type !== 'geometry'">
@@ -145,6 +176,7 @@ export default {
                                         :key="`${property.key}-label`"
                                         :for="`tool-wfs-transaction-form-input-${property.key}`"
                                         class="form-label"
+                                        :class="{'form-label__required': property.required && getInputType(property.type) !== 'checkbox'}"
                                     >
                                         {{ $t(property.label) }}
                                     </label>
@@ -155,17 +187,32 @@ export default {
                                         :type="getInputType(property.type)"
                                         :checked="['true', true].includes(property.value) ? true : false"
                                         class="form-control-checkbox"
-                                        @input="event => setFeatureProperty({key: property.key, type: getInputType(property.type), value: event.target.checked})"
+                                        @input="event => updateFeatureProperty({
+                                            key: property.key,
+                                            type: getInputType(property.type),
+                                            value: event.target.checked
+                                        })"
                                     >
                                     <input
                                         v-else
                                         :id="`tool-wfs-transaction-form-input-${property.key}`"
                                         :key="`${property.key}-input`"
                                         class="form-control"
+                                        :class="{
+                                            'form-control__valid': property.valid === true,
+                                            'form-control__invalid': property.valid === false
+                                        }"
+                                        :step="property.type === 'decimal' ? getDecimalStep(property.type, property.value) : null"
+                                        :title="property.required && !property.valid ? $t(`common:modules.tools.wfsTransaction.mandatoryInputError.${getInputType(property.type)}`): ''"
                                         :type="getInputType(property.type)"
                                         :required="property.required"
                                         :value="property.value"
-                                        @input="event => setFeatureProperty({key: property.key, type: getInputType(property.type), value: event.target.value})"
+                                        @input="event => updateFeatureProperty({
+                                            key: property.key,
+                                            type: getInputType(property.type),
+                                            value: property.type === 'decimal' ? formatDecimalValue(property.type, event.target.value) : event.target.value,
+                                            required: property.required
+                                        })"
                                     >
                                 </template>
                             </template>
@@ -177,6 +224,7 @@ export default {
                                 />
                                 <SimpleButton
                                     :interaction="save"
+                                    :disabled="isFormDisabled"
                                     text="common:modules.tools.wfsTransaction.form.save"
                                     type="button"
                                     class="form-button"
@@ -240,6 +288,19 @@ $margin: 1rem;
             align-self: center;
             margin: 0;
         }
+
+        .form-control.form-control__valid {
+            border: 1px solid green;
+        }
+        .form-control.form-control__invalid {
+            border: 1px solid red;
+        }
+    }
+
+    .form-label__required::after,
+    .form-label-info::before {
+        content: "*";
+        color: red;
     }
 
     .tool-wfs-transaction-form-buttons {