+++ src/modules/tools/wfst/store/mutationsWfst.js
@@ -3,8 +3,11 @@ import initialState from "./stateWfst";
 
 const mutations = {
     ...generateSimpleMutations(initialState),
-    setFeatureProperty ({featureProperties}, {key, value}) {
+
+    setFeatureProperty ({featureProperties}, {key, value, valid, required}) {
         featureProperties.find(property => property.key === key).value = value;
+        featureProperties.find(property => property.key === key).valid = valid;
+        featureProperties.find(property => property.key === key).required = required;
     }
 };