+++ src/modules/tools/wfst/utils/getLayerInformation.js
@@ -10,13 +10,18 @@ function getLayerInformation (layerIds) {
     return layerIds.map(id => {
         const layer = getComponent(id);
 
-        return ["featureNS", "featurePrefix", "featureType", "gfiAttributes", "style", "isSelected", "name", "url", "version"]
-            .reduce((previous, key) => ({...previous, [key]: layer.get(key)}),
-                {
-                    id,
-                    isSecured: layer.get("isSecured") !== undefined ? layer.get("isSecured") : false
-                }
-            );
+        if (layer) {
+            return ["featureNS", "featurePrefix", "featureType", "gfiAttributes", "style", "isSelected", "name", "url", "version"]
+                .reduce((previous, key) => ({...previous, [key]: layer?.get(key)}),
+                    {
+                        id,
+                        isSecured: layer?.get("isSecured") !== undefined ? layer?.get("isSecured") : false
+                    }
+                );
+        }
+
+        return {};
+
     });
 }